nc/requirements.txt
python-telegram-bot==13.13
cachetools==4.2.2